updated POI label text on IMG map renderer to draw after roads

+++ msmap.h
@@ -206,7 +206,7 @@ private:
 	
 	kGUIText m_t;	/* temp used in drawing */
 	void ReadLabel(const char *enc,kGUIString *s);
-	static void DrawLabel(kGUIText *t,int lx,int ly,int lw,int lh,double heading);
+	static void DrawLabel(kGUIText *t,double lx,double ly,double lw,double lh,double heading);
 	void DrawLineLabel(kGUIText *s,int nvert,kGUIDPoint2 *point,int over,bool root);
 	void DrawPolyLabel(kGUIText *s,int nvert,kGUIPoint2 *point);
